src/requirements/_base.txt
djangorestframework==3.*

psycopg2-binary==2.*
